[OG-21] Half code refactor, animation fix
