+++ polymorphia-frontend/components/course/project-section/modals/EventRewardModal.tsx
@@ -15,7 +15,11 @@ export default function AssignmentDetailsModal({
   isError: boolean;
 }) {
   return (
-    <Modal isDataPresented={true} onClosed={onClosed} title="Nagrody">
+    <Modal
+      isDataPresented={data !== undefined}
+      onClosed={onClosed}
+      title="Nagrody"
+    >
       {isError && (
         <div className="gradable-event-section text-xl 2xl:text-2xl">
           Wystąpił błąd przy ładowaniu szczegółów.
